Move the interfaces in their respective apps
