+++ packages/esm-patient-common-lib/src/orders/types/index.ts
@@ -1,4 +1,2 @@
 export * from './order';
-export * from './drug-order';
-export * from './test-order';
 export * from './order-type';
