--- packages/esm-patient-common-lib/src/orders/types/test-order.ts
@@ -1,10 +1,0 @@
-import type { OrderBasketItem } from './order';
-
-export interface TestOrderBasketItem extends OrderBasketItem {
-  testType: {
-    label: string;
-    conceptUuid: string;
-  };
-  orderReason?: string;
-  specimenSource?: string;
-}
