+++ packages/esm-patient-medications-app/src/add-drug-order/drug-search/drug-search.component.tsx
@@ -2,7 +2,8 @@ import React, { useCallback, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Button, Search } from '@carbon/react';
 import { useConfig, useDebounce, ResponsiveWrapper, closeWorkspace, useLayoutType } from '@openmrs/esm-framework';
-import { type DrugOrderBasketItem, launchPatientWorkspace } from '@openmrs/esm-patient-common-lib';
+import { launchPatientWorkspace } from '@openmrs/esm-patient-common-lib';
+import { type DrugOrderBasketItem } from '../../types';
 import { type ConfigObject } from '../../config-schema';
 import OrderBasketSearchResults from './order-basket-search-results.component';
 import styles from './order-basket-search.scss';
