+++ packages/esm-patient-medications-app/src/add-drug-order/drug-search/drug-search.resource.tsx
@@ -1,7 +1,8 @@
 import { useMemo } from 'react';
 import useSWRImmutable from 'swr/immutable';
 import { type FetchResponse, openmrsFetch, restBaseUrl } from '@openmrs/esm-framework';
-import type { Drug, DrugOrderBasketItem, DrugOrderTemplate, OrderTemplate } from '@openmrs/esm-patient-common-lib';
+import type { Drug } from '@openmrs/esm-patient-common-lib';
+import type { DrugOrderBasketItem, DrugOrderTemplate, OrderTemplate } from '../../types';
 
 export interface DrugSearchResult {
   uuid: string;
