+++ packages/esm-patient-medications-app/src/api/api.ts
@@ -2,12 +2,9 @@ import { useCallback, useMemo } from 'react';
 import useSWR, { mutate } from 'swr';
 import useSWRImmutable from 'swr/immutable';
 import { openmrsFetch, restBaseUrl, useConfig, type FetchResponse } from '@openmrs/esm-framework';
-import {
-  type OrderPost,
-  type PatientOrderFetchResponse,
-  type DrugOrderBasketItem,
-} from '@openmrs/esm-patient-common-lib';
+import { type OrderPost, type PatientOrderFetchResponse } from '@openmrs/esm-patient-common-lib';
 import { type ConfigObject } from '../config-schema';
+import type { DrugOrderBasketItem } from '../types';
 
 export const careSettingUuid = '6f0c9a92-6f24-11e3-af88-005056821db0';
