+++ packages/esm-patient-medications-app/src/drug-order-basket-panel/drug-order-basket-panel.extension.tsx
@@ -3,7 +3,8 @@ import { useTranslation } from 'react-i18next';
 import classNames from 'classnames';
 import { Button, Tile } from '@carbon/react';
 import { AddIcon, ChevronDownIcon, ChevronUpIcon, closeWorkspace, useLayoutType } from '@openmrs/esm-framework';
-import { launchPatientWorkspace, useOrderBasket, type DrugOrderBasketItem } from '@openmrs/esm-patient-common-lib';
+import { launchPatientWorkspace, useOrderBasket } from '@openmrs/esm-patient-common-lib';
+import type { DrugOrderBasketItem } from '../types';
 import { prepMedicationOrderPostData } from '../api/api';
 import OrderBasketItemTile from './order-basket-item-tile.component';
 import RxIcon from './rx-icon.component';
