+++ packages/esm-patient-medications-app/src/drug-order-basket-panel/drug-order-basket-panel.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { screen, render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { getByTextWithMarkup } from 'tools';
-import { type DrugOrderBasketItem } from '@openmrs/esm-patient-common-lib';
+import type { DrugOrderBasketItem } from '../types';
 import { mockDrugSearchResultApiData, mockPatientDrugOrdersApiData } from '__mocks__';
 import { getTemplateOrderBasketItem } from '../add-drug-order/drug-search/drug-search.resource';
 import DrugOrderBasketPanel from './drug-order-basket-panel.extension';
