+++ packages/esm-patient-tests-app/src/test-orders/add-test-order/add-test-order.workspace.tsx
@@ -16,7 +16,6 @@ import {
 import {
   type DefaultPatientWorkspaceProps,
   type OrderBasketItem,
-  type TestOrderBasketItem,
   launchPatientWorkspace,
   usePatientChartStore,
 } from '@openmrs/esm-patient-common-lib';
@@ -24,6 +23,7 @@ import { LabOrderForm } from './test-order-form.component';
 import { TestTypeSearch } from './test-type-search.component';
 import styles from './add-test-order.scss';
 import { type ConfigObject } from '../../config-schema';
+import type { TestOrderBasketItem } from '../../types';
 
 export interface AddLabOrderWorkspaceAdditionalProps {
   order?: OrderBasketItem;
