+++ packages/esm-patient-tests-app/src/test-orders/add-test-order/test-order-form.component.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import classNames from 'classnames';
 import {
-  type TestOrderBasketItem,
   type DefaultPatientWorkspaceProps,
   launchPatientWorkspace,
   useOrderBasket,
@@ -29,6 +28,7 @@ import { z } from 'zod';
 import { moduleName } from '@openmrs/esm-patient-chart-app/src/constants';
 import { type ConfigObject } from '../../config-schema';
 import styles from './test-order-form.scss';
+import type { TestOrderBasketItem } from '../../types';
 
 export interface LabOrderFormProps extends DefaultPatientWorkspaceProps {
   initialOrder: TestOrderBasketItem;
