+++ packages/esm-patient-tests-app/src/test-orders/lab-order-basket-panel/lab-order-basket-panel.test.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import userEvent from '@testing-library/user-event';
 import { screen, render } from '@testing-library/react';
-import { useOrderType, type TestOrderBasketItem } from '@openmrs/esm-patient-common-lib';
+import { useOrderType } from '@openmrs/esm-patient-common-lib';
 import LabOrderBasketPanel from './lab-order-basket-panel.extension';
 import { getDefaultsFromConfigSchema, useConfig } from '@openmrs/esm-framework';
 import { type ConfigObject, configSchema } from '../../config-schema';
+import type { TestOrderBasketItem } from '../../types';
 
 const mockUseOrderBasket = jest.fn();
 const mockUseConfig = jest.mocked(useConfig<ConfigObject>);
